Move the trace event store into its own package
